fix for notebooks in sub-directories
I think this is my understanding:
if you add a content with the `--content /absolute/path/fo/my-file.ipynb` then jupyterlite has no idea what the content root is,
so it can't work out the path relative to that,
so the best it can do is expose that file under the path "my-file.ipynb".

So instead, we copy all the notebooks to a directory and link to that.