fix: use all builtin extensions of Markdown when computing TOC
Before, in the example `docs/notebook.ipynb`, there was a markdown block with

```yaml
key: value
...
```

but the default parser doesn't understand that the `# a comment here` is inside a code block and should not be interpreted as a heading.

Now, by enabling the extension, it is interpreted correctly.

I chose to enable all builtin extensions for the highest likelihood that the parsing happens correctly.